pyproject.toml
[project]
name = "adk-mcp-agent"
version = "0.1.0"
description = "ADK agents with MCP code mode pattern"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=0.1.0",
    "fastmcp>=0.1.0",
    "mcp>=1.0.0",
    "python-dotenv>=1.0.0",
    "nest-asyncio>=1.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
